requirements.txt
Django==5.1.5
djangorestframework==3.15.2
celery==5.4.0
redis==5.2.1
fastapi==0.115.7
uvicorn==0.34.0
httpx==0.28.1
groq==0.16.0
django-cors-headers==4.3.1
requests==2.32.3

